Refactor test common methods to test_helpers.go
